+++ src/main/java/dev/lesechko/jdbccrud/repository/LabelRepository.java
@@ -2,6 +2,6 @@ package dev.lesechko.jdbccrud.repository;
 
 import dev.lesechko.jdbccrud.model.Label;
 
-public interface LabelRepository extends GenericRepository<Label, Long> {
+public interface LabelRepository extends GenericRepository<Label, Integer> {
 
 }
